Add schema files, create db entry on click
