+++ src/Clients/Schemas/users.ts
@@ -1,0 +1,11 @@
+export interface User {
+  _id: string,
+  username: string,
+  password: string,
+  firstName: string,
+  lastName: string,
+  email: string,
+  following: string[],
+  followers: string[],
+  savedConcerts: string[],
+}
